Experiment (core) - implementing methods to init, reset and to play only one step

+++ src/core/experimentsmgr.cpp
@@ -99,7 +99,7 @@ void ExperimentsMgr::finished(Experiment* exp)
     }
     emit (statusChanged(exp));
 
-    if (exp->getAutoDelete()) {
+    if (exp->getExpStatus() == Experiment::FINISHED && exp->getAutoDelete()) {
         exp->deleteTrials();
     }
